Tidying up of the xtl_file open statement.
The original code caused problems when using the CRAY compiler,
as the string containing the file name was not being read
correctly. I've add string manipulation to force left-justification
and trim the string (as was being done for the runtime optional
commands), and it now works with CRAY.

I've also updated the error handling for the open file command,
so that the error messages are more informative.